Move Exceptions to own module
