+++ changelog.txt
@@ -1,5 +1,6 @@
 Version 0.17.1 (2022-01-09)
 - Add config option to specify storage directory
+- Move Exceptions to own module
 
 Version 0.17.1 (2022-01-09)
 - Fix naming for multi channel custom entities
