+++ hahomematic/helpers.py
@@ -21,6 +21,7 @@ from hahomematic.const import (
     ATTR_VALUE,
 )
 import hahomematic.devices.entity_definition as hm_entity_definition
+from hahomematic.exceptions import HaHomematicException
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -101,7 +102,7 @@ def check_or_create_directory(directory: str) -> bool:
                 directory,
                 ose.strerror,
             )
-            return False
+            raise HaHomematicException from ose
 
     return True
